feat: done navbar routing
