+++ src/app/(navbar)/community/page.tsx
@@ -1,13 +1,10 @@
-import React from 'react';
+import React from "react";
 import Navbar from "@/app/_components/Navbar";
 
 const Page = () => {
-  
-  const isLoggedIn = true;
-
   return (
     <div>
-      <Navbar isLoggedIn={isLoggedIn} />
+      <Navbar />
       <main className="p-4">
         <h1>Community</h1>
         <p>This is the community page.</p>
