--- src/app/(navbar)/home/merchant/page.tsx
@@ -1,14 +1,0 @@
-import Logout from "@/app/_components/logout";
-import React from "react";
-
-const page = () => {
-  return (
-    <div>
-      <div>
-        <Logout />
-      </div>
-    </div>
-  );
-};
-
-export default page;
