+++ src/app/_components/Navbar.tsx
@@ -1,29 +1,29 @@
 import Link from "next/link";
-import { FC } from "react";
+import React from "react";
+import LoginByRole from "./loginByRole";
+import { getServerAuthSession } from "@/server/auth";
 
-interface NavbarProps {
-  isLoggedIn: boolean;
-  role?: "customer" | "merchant";
-}
+const Navbar = async () => {
+  const session = await getServerAuthSession();
 
-const Navbar: FC<NavbarProps> = ({ isLoggedIn, role = "customer" }) => {
-  const isAdmin = role === "merchant";
-
-  return (
-    <nav className="fixed left-0 right-0 top-0 z-50 m-5 flex h-16 items-center justify-between rounded-2xl border border-transparent bg-[#16a34a] px-6 py-3 text-white">
-      <div className="flex items-center space-x-4">
-        {isAdmin ? (
+  // Merchant
+  if (session?.user.role === "merchant") {
+    return (
+      <nav className="fixed left-0 right-0 top-0 z-50 m-5 flex h-16 items-center justify-between rounded-2xl border border-transparent bg-[#679436] px-6 py-3 text-white">
+        <div className="flex-1 text-center text-lg font-semibold">
+          Nobazir Seller Center
+        </div>
+        <div className="flex space-x-4">
+          <LoginByRole session={session} />
+        </div>
+      </nav>
+    );
+  } else {
+    return (
+      <nav className="fixed left-0 right-0 top-0 z-50 m-5 flex h-16 items-center justify-between rounded-2xl border border-transparent bg-[#679436] px-6 py-3 text-white">
+        <div className="flex items-center space-x-4">
           <div className="flex space-x-4">
-            <Link href="/" className="text-lg hover:text-green-800">
-              Home
-            </Link>
-            <Link href="/product" className="text-lg hover:text-green-800">
-              Product
-            </Link>
-          </div>
-        ) : (
-          <div className="flex space-x-4">
-            <Link href="/" className="text-lg hover:text-green-800">
+            <Link href="/home" className="text-lg hover:text-green-800">
               Home
             </Link>
             <Link href="/leftover" className="text-lg hover:text-green-800">
@@ -36,69 +36,13 @@ const Navbar: FC<NavbarProps> = ({ isLoggedIn, role = "customer" }) => {
               About Us
             </Link>
           </div>
-        )}
-      </div>
-
-      {isAdmin ? (
-        <div className="flex-1 text-center text-lg font-semibold">
-          Nobazir Seller Center
         </div>
-      ) : null}
-
-      <div className="flex space-x-4">
-        {isLoggedIn ? (
-          <>
-            {isAdmin ? (
-              <>
-                <button className="rounded-full p-1 hover:bg-green-700">
-                  <img
-                    src="/navbar/person.svg"
-                    alt="User"
-                    width={32}
-                    height={32}
-                  />
-                </button>
-              </>
-            ) : (
-              <>
-                <button className="rounded-full p-1 hover:bg-green-700">
-                  <img
-                    src="/navbar/cart.svg"
-                    alt="Cart"
-                    width={32}
-                    height={32}
-                  />
-                </button>
-                <button className="rounded-full p-1 hover:bg-green-700">
-                  <img
-                    src="/navbar/person.svg"
-                    alt="User"
-                    width={32}
-                    height={32}
-                  />
-                </button>
-              </>
-            )}
-          </>
-        ) : (
-          <>
-            <Link
-              href="/login"
-              className="flex h-10 w-32 items-center justify-center rounded-full border border-yellow-500 bg-white p-3 text-yellow-500 hover:bg-green-700"
-            >
-              Log In
-            </Link>
-            <Link
-              href="/business-signup"
-              className="flex h-10 items-center justify-center rounded-full bg-yellow-500 p-3 text-white hover:bg-green-700"
-            >
-              Business Sign Up
-            </Link>
-          </>
-        )}
-      </div>
-    </nav>
-  );
+        <div className="flex space-x-4">
+          <LoginByRole session={session} />
+        </div>
+      </nav>
+    );
+  }
 };
 
 export default Navbar;
